Fix the result from .index() replacement
I found when attempting to use the `.index()` replacement that the index was off by 1.
The error was because the `do {}` part of the loop was executing the `i++` before the condition was evaluated meaning the result is always at-least `1`. This is wrong as the first element should have an index of `0`.

Here's a JSFiddle showing the before/after and with the jQuery version as a reference.
https://jsfiddle.net/yck52avu/1/